+++ packages/react-magma-dom/src/components/Combobox/Combobox.tsx
@@ -20,7 +20,9 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
     defaultItems,
     defaultSelectedItem,
     disableCreateItem,
+    errorMessage,
     hasError,
+    helperMessage,
     initialSelectedItem,
     inputStyle,
     isClearable,
@@ -32,9 +34,11 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
     itemListMaxHeight,
     items,
     itemToString,
+    labelPosition,
     labelStyle,
     labelText,
     menuStyle,
+    messageStyle,
     newItemTransform,
     onInputBlur,
     onInputChange,
@@ -247,12 +251,17 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
 
   return (
     <SelectContainer
+      descriptionId={ariaDescribedBy}
+      errorMessage={errorMessage}
       getLabelProps={getLabelProps}
       hasError={hasError}
+      helperMessage={helperMessage}
       isLabelVisuallyHidden={isLabelVisuallyHidden}
       isInverse={isInverse}
+      labelPosition={labelPosition}
       labelStyle={labelStyle}
       labelText={labelText}
+      messageStyle={messageStyle}
     >
       <ComboboxInput
         ariaDescribedBy={ariaDescribedBy}